% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS_dognuts.R
\name{NPS_Dognuts}
\alias{NPS_Dognuts}
\title{Generate a dognuts about recommendation score}
\usage{
NPS_Dognuts(reco)
}
\arguments{
\item{reco}{vector}
}
\value{
plotly figure
}
\description{
Generate a dognuts about recommendation score
}
\examples{
NPS_Dognuts(c(1,2,10,4,3,6,7,8))
}
